+++ src/widgets/clothes/index.ts
@@ -2,5 +2,4 @@ export { getRecommendClothes } from './api/recommend';
 
 export type { RecommendClothesDto } from './model/types';
 
-export { RecommendClothes } from './ui/RecommendClothes/RecommendClothes';
-export { RecommendClothesLoading } from './ui/RecommendClothes/RecommendClothesLoading';
+export { RecommendClothesList } from './ui/RecommendClothesList/RecommendClothesList';
